+++ CHANGES.md
@@ -1,7 +1,8 @@
 # Changes
 
-##### cordova-sqlite-legacy-core 1.0.7-pre4
+##### cordova-sqlite-legacy-core 1.0.7-pre5
 
+- Signal INTERNAL ERROR in case of attempt to reuse db on Windows (should never happen due to workaround solution to BUG 666)
 - SQLITE_DEFAULT_CACHE_SIZE build flag fix for macOS ("osx") and Windows
 
 ###### cordova-sqlite-legacy-express-core 1.0.5